+++ Source/YaOpt/Patches/ThreadSafe/Verse_MapPawns_AssertMainThread.cs
@@ -14,7 +14,7 @@ namespace YaOpt.Patches.ThreadSafe
 
 		static bool Prefix()
 		{
-			return !YaOptGlobal.IsParallelRunningInTick;
+			return !(YaOptGlobal.IsParallelRunningInTick || YaOptGlobal.IsRendering);
 		}
 	}
 }